Fix all compile warnings in comprehensive test suite (15 warnings → 0)
WARNINGS FIXED (15 total):

1. example_chained.c (2 warnings)
   - Line 266, 290: printf format mismatch uint64_t → int
   - Fix: Added explicit (int) cast for bit extraction

2. ml_features.c (1 warning)
   - Line 388: printf %zu with int argument
   - Fix: Changed %zu to %d

3. financial_orderbook.c (4 warnings)
   - Lines 160, 174, 231, 232: incompatible pointer types
   - Fix: Changed anonymous struct to named struct Order

4. varintPacked.h (2 warnings)
   - Line 22: #warning preprocessor directive
   - Fix: Converted #warning to comment (maintains intent, eliminates warning)

5. log_aggregation.c (1 warning)
   - Line 295: printf %zu with int argument
   - Fix: Changed %zu to %d

6. trie_pattern_matcher.c (5 warnings)
   - Lines 332, 386, 387: implicit function declarations
   - Fix: Added #include "varintTagged.h"
   - Lines 1173, 1175: too many arguments for format
   - Fix: Removed unused 0 argument from snprintf

VERIFICATION:
✅ test_all_comprehensive.sh: 41/41 PASSED
✅ Compile warnings: 0 (was 15)
✅ Core library (src/): Still 0 warnings
✅ Examples: Now 0 warnings (was 15)

FILES MODIFIED:
- examples/standalone/example_chained.c
- examples/integration/ml_features.c
- examples/advanced/financial_orderbook.c
- examples/advanced/log_aggregation.c
- examples/advanced/trie_pattern_matcher.c
- src/varintPacked.h

+++ examples/standalone/example_chained.c
@@ -263,7 +263,7 @@ void example_format_comparison() {
 
     printf("Value: %lu (binary: ", value);
     for (int i = 13; i >= 0; i--) {
-        printf("%d", (value >> i) & 1);
+        printf("%d", (int)((value >> i) & 1));
         if (i == 7) printf(" ");
     }
     printf(")\n\n");
@@ -287,7 +287,7 @@ void example_format_comparison() {
     /* SQLite3 varint uses big-endian: first byte has high bits, second byte has low bits */
     uint64_t extracted = ((uint64_t)(buffer[0] & 0x7f) << 7) | (buffer[1] & 0x7f);
     for (int i = 13; i >= 0; i--) {
-        printf("%d", (extracted >> i) & 1);
+        printf("%d", (int)((extracted >> i) & 1));
         if (i == 7) printf(" ");
     }
     printf(" = %lu\n", extracted);
